LayoutTests/http/tests/site-isolation/page-lifecycle/unload-expected.txt

main frame - has 1 onunload handler(s)
Tests that unload is fired when site isolation is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS caughtEvent.name is "unload"
PASS successfullyParsed is true

TEST COMPLETE
